backend/requirements.txt
# 生产依赖
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 数据验证
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# 数据库
pymongo==4.5.0
motor==3.3.2
redis==5.0.1
qdrant-client==1.7.0

# HTTP客户端
httpx==0.26.0
aiohttp==3.9.1

# YAML处理
pyyaml==6.0.1

# 文件监控（热重载）
watchdog==3.0.0

# RSS解析
feedparser==6.0.10

# HTML内容提取
readability-lxml==0.8.1
lxml-html-clean==0.1.1
html2text==2020.1.16
beautifulsoup4==4.12.2

# 日志
python-json-logger==2.0.7

# 工具
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
tenacity==8.2.3

# 测试（生产环境也需要用于健康检查等）
pytest==7.4.4
pytest-asyncio==0.21.1
httpx==0.26.0
